1. Remove test entries 2. replace old routes with new routes
