--- src/views/pages/accounts/create-account/CreateAccount.less
@@ -1,43 +1,0 @@
-/**
- * Copyright 2020 NEM Foundation (https://nem.io)
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-.create-account-wrapper {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  background-color: white;
-  align-items: center;
-  justify-content: center;
-  overflow-y: auto;
-
-  .inner-container {
-    height: 100%;
-    .create-account-box {
-      margin-bottom: 30px;
-      font-size: 35px;
-      color: rgba(0, 0, 0, 0.85);
-      line-height: 40px;
-    }
-
-    .step-bar-container {
-      .step-text-container {
-        span {
-          width: 20%;
-        }
-      }
-    }
-  }
-
-}
