--- src/views/pages/accounts/create-account/finalize/Finalize.less
@@ -1,40 +1,0 @@
-.finish-sec {
-  max-height: calc(100% - 140px);
-
-  .finish-col {
-    display: flex;
-
-    .finish-left {
-      width: 900px;
-    }
-  }
-
-  button {
-    margin-top: 30px;
-    width: 410px;
-    height: 46px;
-  }
-
-  .finish-right {
-    img {
-      width: 300px;
-    }
-  }
-
-  .set-title-tips {
-    margin: 20px 0;
-  }
-
-  .text {
-    padding-bottom: 30px;
-  }
-
-  .text1 {
-    padding-top: 30px;
-  }
-
-  .jump-btn {
-    display: flex;
-    justify-content: space-around;
-  }
-}
