--- src/views/pages/accounts/create-account/generate-mnemonic/GenerateMnemonic.less
@@ -1,31 +1,0 @@
-.step-2-wrapper {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-top: 120px;
-  .progress-bar {
-    text-align: center;
-    margin: 20px auto;
-    font-size: 26px;
-    font-weight: bold;
-    color: rgba(0, 0, 0, 0.85);
-    padding-bottom: 20px;
-  }
-  .ivu-progress {
-    width: 500px;
-    .ivu-progress-text-inner {
-      color: #6401ee;
-    }
-  }
-  .ivu-progress-text-inner {
-    position: absolute;
-    font-size: 60px;
-    font-weight:400;
-    bottom: 100px;
-    right: 50%;
-    text-align: center;
-    color: rgba(82,0,198,1);
-  }
-}
