--- src/views/pages/accounts/create-account/show-mnemonic/ShowMnemonic.less
@@ -1,79 +1,0 @@
-/**
- * Copyright 2020 NEM Foundation (https://nem.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@import "../../../../resources/css/variables.less";
-
-.create-mnemonic-sec {
-  position: relative;
-
-  .create-mnemonic-col {
-    display: flex;
-
-    .create-mnemonic-left {
-      width: 950px;
-
-      .show-mnemonic {
-        width: 950px;
-        height: 260px;
-        display: flex;
-        align-items: center;
-        flex-direction: column;
-
-        img {
-          width: 60px;
-          margin-bottom: 10px;
-        }
-      }
-
-      .button-container {
-        width: 100%;
-        clear: both;
-      }
-    }
-
-    .create-mnemonic-right {
-      padding-left: 80px;
-      width: 510px;
-    }
-  }
-
-  .mnemonic-container {
-    padding-top: 5px;
-    width: 950px;
-    height: 260px;
-    border: 1px solid rgba(51, 51, 51, 0.2);
-    border-radius: 5px;
-    margin-bottom: 20px;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-  }
-
-  .mnemonic-list {
-    padding: 10px;
-
-    span {
-      display: inline-block;
-      background-color: #6401ee;
-      padding: 5px 10px;
-      border-radius: 4px;
-      margin: 5px;
-      font-size: @smallFont;
-      color: rgba(255, 255, 255, 0.9);
-    }
-
-  }
-}
