--- src/views/pages/accounts/create-account/verify-mnemonic/VerifyMnemonic.less
@@ -1,63 +1,0 @@
-/**
- * Copyright 2020 NEM Foundation (https://nem.io)
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- 
-.check-mnemonic-sec {
-  position: relative;
-
-  .skip {
-    float: right;
-    margin-right: 700px;
-    font-size: 20px;
-    color: #5200c6;
-    width: 100px;
-    height: 20px;
-  }
-
-  .check-mnemonic-col {
-    display: flex;
-
-    .check-mnemonic-left {
-      width: 1000px;
-
-      .show-mnemonic {
-        width: 1300px;
-        height: 190px;
-        border: 1px solid rgba(51, 51, 51, 0.2);
-        border-radius: 5px;
-        margin-bottom: 20px;
-
-        img {
-          margin-top: 30px;
-          margin-left: 240px;
-        }
-      }
-    }
-
-    .check-mnemonic-right {
-      width: 530px;
-
-    }
-  }
-
-  .buttons {
-    width: 100%;
-  }
-
-  .ivu-btn {
-    padding: 0 150px;
-  }
-
-}
