--- src/views/pages/accounts/import-account/ImportAccount.less
@@ -1,25 +1,0 @@
-.import-account-wrapper {
-  width: 100%;
-  height: 100%;
-  background-color: white;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  .create-account-box {
-    margin-bottom: 40px;
-    font-size: 40px;
-    font-weight: 400;
-    color: rgba(0, 0, 0, 0.85);
-    line-height: 40px;
-  }
-
-  .step-bar-container {
-    .step-text-container {
-      span {
-        width: 25%;
-      }
-    }
-  }
-}
