--- src/views/pages/accounts/import-account/finalize/Finalize.less
@@ -1,35 +1,0 @@
-.finish-sec {
-  max-height: calc(100% - 150px);
-
-  .finish-col {
-    display: flex;
-
-    .finish-left {
-      width: 900px;
-    }
-  }
-
-  button {
-    margin-top: 30px;
-    width: 410px;
-    height: 50px;
-  }
-
-  .finish-right {
-    img {
-      width: 300px;
-    }
-  }
-
-  .set-title-tips {
-    margin: 10px 0;
-  }
-
-  .text1 {
-    padding-top: 30px;
-  }
-
-  .text {
-    padding-bottom: 30px;
-  }
-}
