--- src/views/pages/accounts/import-account/finalize/FinalizeTs.ts
@@ -1,6 +1,0 @@
-import {Vue, Component} from 'vue-property-decorator'
-
-@Component
-export default class FinalizeTs extends Vue {
-
-}
