--- src/views/pages/accounts/import-account/import-mnemonic/ImportMnemonic.less
@@ -1,42 +1,0 @@
-.create-mnemonic-sec {
-  position: relative;
-
-  .button-container {
-    display: inline;
-  }
-
-  .create-mnemonic-col {
-    display: flex;
-
-    .create-mnemonic-left {
-      width: 950px;
-
-      .show-mnemonic {
-        width: 930px;
-        height: 260px;
-        padding: 20px;
-        border: 1px solid rgba(51, 51, 51, 0.2);
-        border-radius: 5px;
-        margin-bottom: 20px;
-      }
-
-    }
-
-    .create-mnemonic-right {
-      padding-left: 80px;
-      width: 560px;
-
-    }
-  }
-
-  .button-container {
-    width: 400px;
-
-    button {
-      margin-top: 20px;
-      width: 347px;
-      height: 50px;
-      margin-right: 16px;
-    }
-  }
-}
