--- src/views/pages/accounts/import-account/wallet-selection/WalletSelection.less
@@ -1,173 +1,0 @@
-.choose-wallet-sec {
-  height: calc(100% - 160px);
-  .left-container {
-    height: calc(100% - 10px);
-
-    .dialog-title {
-      font-size: 28px;
-      color: rgba(51, 51, 51);
-      padding: 0 0 10px 0;
-    }
-
-    .dialog-sub-tips {
-      font-size: 22px;
-      color: rgba(51, 51, 51, 1);
-      padding: 10px 0 20px 0;
-      font-weight: bolder;
-    }
-
-    .choose-hd-path {
-      border: 1px solid rgba(224, 223, 229);
-      width: 700px;
-      height: calc(100% - 60px);
-      padding: 10px 0;
-    }
-
-    .hd-path {
-      font-size: 20px;
-      padding: 10px 30px;
-      border-bottom: 1px solid rgba(224, 223, 229);
-
-      .hd-path-title {
-        margin-right: 40px;
-        color: rgba(51, 51, 51);
-      }
-
-      .hd-path-type,
-      .hd-path-value {
-        color: rgba(136, 136, 136);
-        line-height: 17px;
-      }
-    }
-
-  }
-
-  .right-container {
-    position: absolute;
-    top: 262px;
-    right: 210px;
-    min-height: 500px;
-
-    .address-list-container {
-      border: 1px solid rgba(224, 223, 229);
-      width: 700px;
-      height: 470px;
-      padding: 10px 0 5px 0;
-      font-size: 16px;
-      color: rgba(51, 51, 51);
-    }
-
-    .address-id {
-      margin-right: 30px;
-    }
-
-    .right-container-title {
-      font-size: 0.22rem;
-      color: #333333;
-      padding: 0.1rem 0 0.2rem 0;
-      font-weight: bolder;
-    }
-
-    .remove-icon {
-      width: 50px;
-
-      img {
-        width: 21px;
-        height: 21px;
-      }
-
-    }
-  }
-
-  .address-list {
-    font-size: 16px;
-    height: 300px;
-
-    .scroll {
-      max-height: 520px;
-    }
-
-    .table-item,
-    .table-title {
-      padding: 5px 14px;
-      color: rgba(150, 150, 153);
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-    }
-
-    .table-title .address-id {
-      position: relative;
-      left: 20px;
-    }
-
-
-    .table-item {
-      padding: 5px 0;
-      border-bottom: 1px solid #E0DFE5;
-    }
-
-    .table-item:hover {
-      color: #1FB6AC;
-      background-color: rgba(38, 180, 176, 0.1);
-    }
-
-    .table-item:last-of-type {
-      border-bottom: 1px solid transparent;
-    }
-
-    .address-id {
-      width: 30px;
-      position: relative;
-      left: 30px;
-      font-size: 20px;
-
-      img {
-        width: 20px;
-        height: 20px;
-        position: relative;
-        left: 20px;
-      }
-    }
-
-    .address-value {
-      width: 450px;
-      font-size: 20px;
-    }
-
-    .address-balance {
-      font-size: 20px;
-      width: 120px;
-    }
-  }
-
-  .button-container {
-    display: flex;
-    padding-top: 20px;
-    align-items: center;
-
-    button {
-      margin-top: 20px;
-      width: 407px;
-      height: 50px;
-    }
-  }
-
-  .selection-container {
-    display: flex;
-    align-items: center;
-    justify-content: space-around;
-    font-size: 20px;
-
-    .network-type {
-      display: inline-block;
-      height: 30px;
-      line-height: 30px;
-      width: 40%;
-      text-align: center;
-      border: 1px solid #ccc;
-    }
-  }
-
-
-}
